Implement lastModified check for sources
The annotation processor should be able to skip the annotation
processing if the sources haven't changed since the last processing.
This will speed up project builds considerably since for example the
JPA metamodel generator will not run if the entities haven't been
modified, which will save the compiler of having to recompile your
whole module.

Using the option `skipWithoutSourceChanges` you can enable this
behavior. By default it is turned off to maintain the current behavior.

This change doesn't track individual files, as there need not be a 1-1
mapping between the origin of the annotation processor and the
generated sources. The plugin rather determines from all source
locations what the most recent last modified time is, and does the same
for all the files in the output folder.

This cuts down rebuild times on my current project by a half or so
(going from over 2 minutes to just 1 minute).